backend/requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.0.1